refactor(provider): split providers into separate files for maintainability
closes #812
