+++ commitizen/providers/composer_provider.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+
+from commitizen.providers.base_provider import JsonProvider
+
+
+class ComposerProvider(JsonProvider):
+    """
+    Composer version management
+    """
+
+    filename = "composer.json"
+    indent = 4
